Optimize strcmp in xfce-rc-simple.
Because we heavily rely on strcmp in this code, comparing
the first character gives a nice speedup.

Test app read 100 groups, with each 200 keys. When the
group and key names started with the same character there
was a 3% slowdown, when the first character was 'unique'
(in a 0-9 range) a 35% speedup when looking at the total
function cost of callgrind.