Fix abicheck to handle ppc64 architecture
According to Alan Modra, the PowerPC64 ABI defines function symbols on their
OPD entry, in the .opd section which is part of the data segment; so
functions on this architecture have symbol type D in nm.

The adjusted regex is slightly more accurate than the old one, which matched
anything containing " T", "R", or "G ".  Using a bracket expression seems to
better match the likely intent of this regex.