Add support for GOBJECT-INTROSPECTION (Bug #12159)
GOjbect Introspection allows various other languages to automatically
generate bindings they can use. For details see:
https://wiki.gnome.org/action/show/Projects/GObjectIntrospection

To support this, we need to add a gobject and introspection
dependency. The introspection one is optional and it is what
determines if the libxfce4util-1.0 gir is created.
Additionally, the Kiosk was converted to a GObject to better
work with this method of use, but retains the exact same
functionality externally.

Signed-off-by: Simon Steinbeiss <simon.steinbeiss@elfenbeinturm.at>