session-management: Try to create D-BUS proxy synchronously
The problem arises, for example, with xfce4-power-manager-settings,
which starts xfce4-power-manager if it's not already running, and then
calls org.xfce.Power.Manager.GetConfig, which in turn queries
XfceSystemd for sleep, shutdown, etc. capabilities. The proxy may not
yet have been created in this case.

Generally speaking, it's probably best if the returned object is
immediately (actually) usable, if possible, even if it means blocking
the time needed for this.

Fixes: c11acc334a00d65fde8211a1ed525cd76a0c88ec