xfce-rc: Fix support for the LANGUAGE environment variable
When LANGUAGE contained more than one language code and there was a
translation to the first language in a .desktop file, a translation to
the second language was returned in the case when the line for the
second language was above the line for the first one. Now translation
for the first language is returned regardless of the line order. The
`entry != NULL` condition is inverted to make the diff smaller and
reduce the amount of indentation.

Fixes: 3a279d2b5fd5315e0273bf2ced12dcbb9ab9de46