meson-build: Let pkg.generate() fill in required libs
It completes what we put in 'requires' from the dependencies declared in
library() anyway, by filling in the 'requires_private' field, whose use
is recommended by default. So we probably don't need to fill this in in
general: it's one less thing to maintain, and the dependencies we've
declared public so far should probably have been declared private
instead.

See-also:
https://mesonbuild.com/Pkgconfig-module.html
https://people.freedesktop.org/~dbn/pkg-config-guide.html
https://stackoverflow.com/a/61674208